refactor(simulation): derive SimulationConfig.model from llm_config
Review feedback. `model` was a second field set beside `llm_config` at each construction site, documented as always equal to `llm_config.model` but held there by discipline — a rule a third site can forget. It is now a property, so there is nothing to keep in step.

Moves the `generate_structured(config=)` merge tests to tests/common/test_structured_output_config.py, next to the function they cover rather than under the simulation surface that happens to have prompted them.
