+++ src/evaluatorq/simulation/api.py
@@ -567,7 +567,6 @@ async def _simulate_run(
                         memory_entity_id=memory_entity_id,
                         previous_run=previous_run,
                         max_turns=max_turns,
-                        model=resolved_llm_config.model,
                         llm_config=resolved_llm_config,
                         evaluator_names=evaluator_names,
                         scoring=scoring,
@@ -1015,7 +1014,6 @@ async def _generate_and_simulate_run(
                             dataset_id=None,
                             memory_entity_id=memory_entity_id,
                             max_turns=max_turns,
-                            model=resolved_llm_config.model,
                             llm_config=resolved_llm_config,
                             evaluator_names=evaluator_names,
                             scoring=scoring,
