+++ tests/simulation/test_sim_llm_config.py
@@ -12,7 +12,6 @@ from typing import Any
 from unittest.mock import AsyncMock, MagicMock
 
 import pytest
-from pydantic import BaseModel
 
 from evaluatorq.contracts import LLMCallConfig
 from evaluatorq.simulation.agents.base import AgentConfig
@@ -26,10 +25,6 @@ from evaluatorq.simulation.generators import (
 )
 from evaluatorq.simulation.runner.simulation import SimulationRunner
 from evaluatorq.simulation.types import DEFAULT_MODEL, CommunicationStyle, Persona, Scenario
-
-
-class _Answer(BaseModel):
-    answer: str
 
 
 def _persona() -> Persona:
@@ -230,60 +225,6 @@ async def test_executive_summary_sends_no_temperature_when_unset(monkeypatch: py
 
 
 @pytest.mark.asyncio
-async def test_generate_structured_reads_only_set_config_fields(monkeypatch: pytest.MonkeyPatch) -> None:
-    """The one merge every generator funnels through — and the only test of it."""
-    import evaluatorq.common.structured_output as mod
-
-    seen: dict[str, Any] = {}
-
-    async def fake_parse(**kwargs: Any) -> Any:
-        seen.update(kwargs)
-        raise RuntimeError('stop after the params are built')
-
-    monkeypatch.setattr(mod, 'execute_chat_parse', fake_parse)
-    cfg = LLMCallConfig(model='config/model', temperature=0.7, extra_body={'from': 'config'})
-    with pytest.raises(Exception, match='stop after the params are built'):
-        await mod.generate_structured(
-            MagicMock(),
-            model='explicit/model',
-            messages=[{'role': 'user', 'content': 'hi'}],
-            response_format=_Answer,
-            max_tokens=64,
-            label='test',
-            config=cfg,
-        )
-    assert seen['temperature'] == 0.7
-    assert seen['extra_body'] == {'from': 'config'}
-    # config.model is deliberately not read — the call site stays the authority.
-    assert seen['model'] == 'explicit/model'
-
-
-@pytest.mark.asyncio
-async def test_generate_structured_explicit_keyword_beats_the_config(monkeypatch: pytest.MonkeyPatch) -> None:
-    import evaluatorq.common.structured_output as mod
-
-    seen: dict[str, Any] = {}
-
-    async def fake_parse(**kwargs: Any) -> Any:
-        seen.update(kwargs)
-        raise RuntimeError('stop')
-
-    monkeypatch.setattr(mod, 'execute_chat_parse', fake_parse)
-    with pytest.raises(Exception, match='stop'):
-        await mod.generate_structured(
-            MagicMock(),
-            model='m',
-            messages=[{'role': 'user', 'content': 'hi'}],
-            response_format=_Answer,
-            max_tokens=64,
-            label='test',
-            temperature=0.0,
-            config=LLMCallConfig(model='m', temperature=0.7),
-        )
-    assert seen['temperature'] == 0.0
-
-
-@pytest.mark.asyncio
 @pytest.mark.parametrize('entry', ['summarize_conversations', 'datapoints_from_traces', 'extend_from_traces'])
 async def test_trace_helpers_use_the_config_model(monkeypatch: pytest.MonkeyPatch, entry: str) -> None:
     """`generate_structured` ignores `config.model`, so the entry point must resolve it.
@@ -344,62 +285,6 @@ def test_generators_use_the_config_client_without_env_credentials(
     gen = cls(config=LLMCallConfig(model='m', client=mine))
     assert gen._client is mine
     assert gen._client_owned is False
-
-
-@pytest.mark.asyncio
-async def test_generate_structured_explicit_none_temperature_beats_the_config(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """`None` means "omit the parameter" — a real value, not "caller said nothing"."""
-    import evaluatorq.common.structured_output as mod
-
-    seen: dict[str, Any] = {}
-
-    async def fake_parse(**kwargs: Any) -> Any:
-        seen.update(kwargs)
-        raise RuntimeError('stop')
-
-    monkeypatch.setattr(mod, 'execute_chat_parse', fake_parse)
-    with pytest.raises(Exception, match='stop'):
-        await mod.generate_structured(
-            MagicMock(),
-            model='m',
-            messages=[{'role': 'user', 'content': 'hi'}],
-            response_format=_Answer,
-            max_tokens=64,
-            label='test',
-            temperature=None,
-            config=LLMCallConfig(model='m', temperature=0.7),
-        )
-    assert seen['temperature'] is None
-
-
-@pytest.mark.asyncio
-async def test_generate_structured_explicit_default_timeout_beats_the_config(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Passing the default value on purpose is still passing it."""
-    import evaluatorq.common.structured_output as mod
-
-    seen: dict[str, Any] = {}
-
-    async def fake_parse(**kwargs: Any) -> Any:
-        seen.update(kwargs)
-        raise RuntimeError('stop')
-
-    monkeypatch.setattr(mod, 'execute_chat_parse', fake_parse)
-    with pytest.raises(Exception, match='stop'):
-        await mod.generate_structured(
-            MagicMock(),
-            model='m',
-            messages=[{'role': 'user', 'content': 'hi'}],
-            response_format=_Answer,
-            max_tokens=64,
-            label='test',
-            timeout_s=mod._STRUCTURED_TIMEOUT_S,
-            config=LLMCallConfig(model='m', timeout_ms=1000),
-        )
-    assert seen['timeout_s'] == mod._STRUCTURED_TIMEOUT_S
 
 
 @pytest.mark.asyncio
